Share SOTA evidence run utilities
